Add files via upload
Run the script on a Linux or Unix system where Bioconda or seqtk is installed. If you want to know the percentage of nucleotide marked as "N" in human chromosome 22, then run it like so: ./GetUnknownPercent.sh -c 22.  This script will download chromosome 22 sequence from the GRCh38 genomic build and will print the total count of nucleotides in the chosen chromosome, the count of nucleotides marked "N" and its percentage relative to the total count.